arm64: lse: Omit LL/SC alternatives patching
CAF appears to have messed with some code in this kernel related to LSE
atomics and/or alternatives, causing the combined LSE + LL/SC
out-of-line calling code to be too big for its section when compiling
kernel/locking/spinlock.c. This causes gas to fail with a confusing error:

    /tmp/spinlock-79343b.s: Assembler messages:
    /tmp/spinlock-79343b.s:61: Error: attempt to move .org backwards
    /tmp/spinlock-79343b.s:157: Error: attempt to move .org backwards

Clang's integrated assembler is more verbose and provides a more helpful
error that points to the alternatives code as being the culprit:

    In file included from ../kernel/locking/spinlock.c:20:
    In file included from ../include/linux/spinlock.h:88:
    ../arch/arm64/include/asm/spinlock.h:76:15: error: invalid .org offset '56' (at offset '60')
            asm volatile(ARM64_LSE_ATOMIC_INSN(
                         ^
    ../arch/arm64/include/asm/lse.h:36:2: note: expanded from macro 'ARM64_LSE_ATOMIC_INSN'
            ALTERNATIVE(llsc, lse, ARM64_HAS_LSE_ATOMICS)
            ^
    ../arch/arm64/include/asm/alternative.h:281:2: note: expanded from macro 'ALTERNATIVE'
            _ALTERNATIVE_CFG(oldinstr, newinstr, __VA_ARGS__, 1)
            ^
    ../arch/arm64/include/asm/alternative.h:83:2: note: expanded from macro '_ALTERNATIVE_CFG'
            __ALTERNATIVE_CFG(oldinstr, newinstr, feature, IS_ENABLED(cfg), 0)
            ^
    ../arch/arm64/include/asm/alternative.h:73:16: note: expanded from macro '__ALTERNATIVE_CFG'
            ".popsection\n\t"                                               \
                          ^
    <inline asm>:35:7: note: instantiated into assembly here
            .org    . - (664b-663b) + (662b-661b)
                    ^

Omitting the alternatives code indeed reduces the size enough to make
everything compile successfully. We don't need the patching anyway
because we will only enable CONFIG_ARM64_LSE_ATOMICS when the target CPU
is known to support LSE atomics with 100% certainty, so kill all the
dynamic out-of-line LL/SC patching code.

This change also has the side-effect of reducing the I-cache footprint
of these critical locking and atomic paths, which can reduce cache
thrashing and increase overall performance.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>